refactor: rename model class path and introduce EyeWidthRegressor - Update class path in training configuration and add new EyeWidthRegressor model implementation for enhanced eye width prediction capabilities.
